refactor(server): simplify and consolidate GitHub App logic into a single file
