+++ shared-internal/src/main/kotlin/io/github/typesafegithub/workflows/shared/internal/GithubUtils.kt
@@ -1,10 +1,8 @@
 package io.github.typesafegithub.workflows.shared.internal
 
-import io.github.typesafegithub.workflows.shared.internal.app.GitHubApp
 import kotlinx.coroutines.runBlocking
 
-private val githubApp = GitHubApp()
 
-fun getAppAccessToken(): String = runBlocking { githubApp.accessToken() }
+fun getAppAccessToken(): String = runBlocking { getInstallationAccessToken() }
 
 fun getAppAccessTokenOrNull(): String? = runCatching { getAppAccessToken() }.getOrNull()
